Factor out `DnsTrace`, `ErrorPolicyTrace` & `MuxTrace`  to their own modules
